chore: rename configuration.ts to configuration-wire-types.ts
