+++ src/client/eppo-precomputed-client.spec.ts
@@ -11,9 +11,9 @@ import {
   ensureContextualSubjectAttributes,
   ensureNonContextualSubjectAttributes,
 } from '../attributes';
-import { IPrecomputedConfigurationResponse } from '../configuration';
 import { IConfigurationStore } from '../configuration-store/configuration-store';
 import { MemoryOnlyConfigurationStore } from '../configuration-store/memory.store';
+import { IPrecomputedConfigurationResponse } from '../configuration-wire-types';
 import { DEFAULT_POLL_INTERVAL_MS, MAX_EVENT_QUEUE_SIZE, POLL_JITTER_PCT } from '../constants';
 import FetchHttpClient from '../http-client';
 import {
