+++ src/configuration-wire-types.ts
